test of the toByte() of the mutations

+++ src/main/java/brain/mutation/MutationAdditionLink.java
@@ -18,16 +18,16 @@ public class MutationAdditionLink extends Mutation {
 	 * @param factor the factor of the link
 	 */
 	public MutationAdditionLink(
-			short originArray, 
-			short originPosition, 
-			short targetArray, 
-			short targetPosition,
+			int originArray, 
+			int originPosition, 
+			int targetArray, 
+			int targetPosition,
 			float factor
 			) {
-		this.originArray = originArray;
-		this.originPosition = originPosition;
-		this.targetArray = targetArray;
-		this.targetPosition = targetPosition;
+		this.originArray = (short) originArray;
+		this.originPosition = (short) originPosition;
+		this.targetArray = (short) targetArray;
+		this.targetPosition = (short) targetPosition;
 		this.factor = factor;
 	}
